+++ src/block_io/dynamic.rs
@@ -257,7 +257,7 @@ impl<'a> DynamicBlockIo<'a> {
     }
 
     /// Get block size from BAT
-    pub fn block_size(&self) -> u32 {
+    pub fn block_size(&self) -> u64 {
         self.bat.block_size
     }
