+++ src/block_io/fixed.rs
@@ -96,7 +96,7 @@ impl<'a> FixedBlockIo<'a> {
     }
 
     /// Get block size from BAT
-    pub fn block_size(&self) -> u32 {
+    pub fn block_size(&self) -> u64 {
         self.bat.block_size
     }
 }
